Remove 'node set-voting-delegate'
